+++ ChangeLog.md
@@ -5,6 +5,9 @@ XP Framework Core ChangeLog
 
 ### Features
 
+* Added new method `lang.Process::running()` to test whether a process
+  is running
+  (@thekid)
 * Merged PR #163: Make adding property file expansions easy to use
   (@thekid)